chore(tanstack-start): fix and test type definitions
