+++ packages/tanstack-start/README.md
@@ -1,11 +1,11 @@
 <div align="center">
 
-<h1><b>@aura-stack/integration</b></h1>
+<h1><b>@aura-stack/tanstack-start</b></h1>
 
 **Integration authentication library for the Aura Stack ecosystem**
 
-[![npm version](https://img.shields.io/npm/v/@aura-stack/integration.svg)](https://www.npmjs.com/package/@aura-stack/integration)
-[![JSR version](https://jsr.io/badges/@aura-stack/integration)](https://jsr.io/@aura-stack/integration)
+[![npm version](https://img.shields.io/npm/v/@aura-stack/tanstack-start.svg)](https://www.npmjs.com/package/@aura-stack/tanstack-start)
+[![JSR version](https://jsr.io/badges/@aura-stack/tanstack-start)](https://jsr.io/@aura-stack/tanstack-start)
 [![License: MIT](https://img.shields.io/badge/License-MIT-blue.svg)](https://opensource.org/licenses/MIT)
 
 [Official Docs](https://aura-stack-auth.vercel.app/docs) · [Core Package Docs](https://aura-stack-auth.vercel.app/docs/packages/core)
@@ -17,7 +17,7 @@
 ## Installation
 
 ```bash
-pnpm add @aura-stack/integration
+pnpm add @aura-stack/tanstack-start
 ```
 
 ## Documentation
