+++ packages/tanstack-start/tsconfig.json
@@ -1,10 +1,11 @@
 {
-  "extends": "@aura-stack/tsconfig/tsconfig.base.json",
+  "extends": "@aura-stack/tsconfig/tanstack.json",
   "compilerOptions": {
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./src/*"],
+      "@test/*": ["./test/*"]
     }
   },
-  "include": ["src"],
+  "include": ["src", "test"],
   "exclude": ["dist", "node_modules"]
 }
