+++ packages/tanstack-start/vitest.config.ts
@@ -1,0 +1,18 @@
+import path from "path"
+import { defineConfig } from "vitest/config"
+
+export default defineConfig({
+    test: {
+        include: ["test/**/*.test.tsx", "test/**/*.test.ts"],
+        environment: "jsdom",
+        globals: true,
+        clearMocks: true,
+        restoreMocks: true,
+    },
+    resolve: {
+        alias: {
+            "@": path.resolve(__dirname, "./src"),
+            "@test": path.resolve(__dirname, "./test"),
+        },
+    },
+})
